Added Character to map. Implemented keyboard and gamepad input.  Implemented moving animations for character.
BP_ShooterCharacter added as the default pawn for the BP_ShooterGameMode.
Input implemented in ShooterCharacter.cpp, and in Project Settings.
Animations: Blend space BS_Locomation created to determine which animation to use based on speed and angle of the character. ABP_ShooterCharacter blueprint to actually receive the speed and angle from our character as input, then change the skeletal mesh accordingly.
Applied a collision fix in an asset of the Modular SciFi Season 1 Pack.