Implemented enemy AI
BT_EnemyAI is the behavior tree for the AI. BB_EnemyAI contains the blackboard with the variables we want to manipulate and check in the behavior tree.
BP_ShooterAIController is where we attach BT_EnemyAI. The AIController is then attached to BP_ShooterCharacter.
Added some custom BTServices and BTTasks to update the Blackboard during the flow of the behavior tree.